+++ src/test/java/com/global/api/tests/network/vaps/VapsDebitTests.java
@@ -75,8 +75,6 @@ public class VapsDebitTests {
         acceptorConfig.setSupportsEmvPin(true);
         acceptorConfig.setPinlessDebit(false);
         acceptorConfig.setSupportWexAdditionalProducts(true);
-        acceptorConfig.setSupportVisaFleet2dot0(PurchaseType.NOVISAFLEET2DOT0);
-
 
         // gateway config
         config = new NetworkGatewayConfig();